pyproject.toml
[project]
name = "watcheye"
version = "0.1.0"
description = "Social media content watcheye for Doughnut brand"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "apify-client>=1.6",
    "streamlit>=1.30",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "rich>=13.0",
    "google-genai>=1.0",
]

[project.scripts]
watcheye = "watcheye.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/watcheye"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
